+++ cli/cli.js
@@ -77,23 +77,19 @@ class DjinnJS {
             /** TODO: Add chalk */
             await this.preflightCheck();
             await this.createTempDirectory();
-            await this.createInjectionsDirectory();
             await this.resetOutputDirectories();
             await this.createOutputDirectories();
 
             if (!this.silent) {
                 spinner.text = "Scrubbing JavaScript imports";
             }
-            await this.dependencyInjection();
             await this.scrubScripts();
             await this.injectConfig();
-            await this.injectRuntime();
 
             if (!this.silent) {
                 spinner.text = "Minifying JavaScript";
             }
             await this.minifyScript();
-            await this.relocateServiceWorker();
 
             if (!this.silent) {
                 spinner.text = "Relocating WebAssembly files";
@@ -114,7 +110,6 @@ class DjinnJS {
                 spinner.text = "Cleaning up DjinnJS temporary files";
             }
             await this.cleanup();
-            await this.cachebust();
 
             /** Exit the process when everything runs successfully */
             spinner.succeed("DjinnJS");
@@ -144,29 +139,6 @@ class DjinnJS {
                 }
             });
         }
-    }
-
-    cachebust() {
-        return new Promise((resolve, reject) => {
-            const output = path.resolve(cwd, this.config.publicDir, "resources-cachebust.json");
-            if (fs.existsSync(output)) {
-                fs.unlinkSync(output);
-            }
-
-            const cachebustFile = path.join(__dirname, "resources-cachebust.json");
-            fs.readFile(cachebustFile, (errror, buffer) => {
-                if (errror) {
-                    reject(errror);
-                }
-                let data = buffer.toString().replace("REPLACE_WITH_TIMESTAMP", `${Date.now()}`);
-                fs.writeFile(output, data, errror => {
-                    if (errror) {
-                        reject(errror);
-                    }
-                    resolve();
-                });
-            });
-        });
     }
 
     generateNoScriptCSS() {
@@ -215,44 +187,6 @@ class DjinnJS {
         });
     }
 
-    relocateServiceWorker() {
-        return new Promise((resolve, reject) => {
-            const publicPath = path.resolve(cwd, this.config.publicDir);
-            const assetPath = path.resolve(publicPath, this.config.outDir);
-            if (this.config.serviceWorker) {
-                fs.copyFileSync(`${assetPath}/${this.config.serviceWorker}.mjs`, `${publicPath}/service-worker.js`);
-
-                fs.unlinkSync(`${assetPath}/offline-backup.mjs`);
-                fs.unlinkSync(`${assetPath}/offline-first.mjs`);
-                fs.unlinkSync(`${assetPath}/resources-only.mjs`);
-
-                fs.readFile(`${publicPath}/service-worker.js`, (error, buffer) => {
-                    if (error) {
-                        reject(error);
-                    }
-
-                    let data = buffer.toString();
-                    data = data.replace("REPLACE_WITH_NO_CACHE_PATTERN", this.config.noCachePattern);
-                    data = data.replace("REPLACE_WITH_CACHEBUST_URL", this.config.cachebustURL);
-                    data = data.replace("REPLACE_WITH_RESOURCE_PATTERN", this.config.resourcePattern);
-                    data = data.replace("REPLACE_WITH_PRECACHE_URL", this.config.precacheURL);
-
-                    fs.writeFile(`${publicPath}/service-worker.js`, data, error => {
-                        if (error) {
-                            reject(error);
-                        }
-                        resolve();
-                    });
-                });
-            } else {
-                fs.unlinkSync(`${assetPath}/offline-backup.mjs`);
-                fs.unlinkSync(`${assetPath}/offline-first.mjs`);
-                fs.unlinkSync(`${assetPath}/resources-only.mjs`);
-                resolve();
-            }
-        });
-    }
-
     injectConfig() {
         return new Promise((resolve, reject) => {
             const configFile = path.join(__dirname, "temp", "config.js");
@@ -262,12 +196,7 @@ class DjinnJS {
                 }
                 let data = buffer.toString();
                 data = data.replace("REPLACE_WITH_OUTPUT_DIR_NAME", this.config.outDir);
-                data = data.replace("REPLACE_WITH_GTAG_ID", this.config.gtagId);
-                data = data.replace("REPLACE_WITH_PREFETCH_STATUS", this.config.prefetching);
-                data = data.replace('"REPLACE_WITH_FOLLOW_REDIRECT_STATUS"', this.config.followRedirects);
-                data = data.replace('"REPLACE_WITH_PAGE_JUMP_OFFSET"', this.config.pageJumpOffset);
                 data = data.replace("REPLACE_WITH_MINIMUM_CONNECTION", this.config.minimumConnection);
-                data = data.replace('"REPLACE_WITH_USE_SERVICE_WORKER"', this.config.serviceWorker ? true : false);
                 fs.writeFile(configFile, data, error => {
                     if (error) {
                         reject(error);
@@ -276,81 +205,6 @@ class DjinnJS {
                 });
             });
         });
-    }
-
-    injectRuntime() {
-        return new Promise((resolve, reject) => {
-            const runtimeFile = path.join(__dirname, "temp", "runtime.js");
-            fs.readFile(runtimeFile, (error, buffer) => {
-                if (error) {
-                    reject(error);
-                }
-                let data = buffer.toString();
-
-                if (this.config.pjax) {
-                    const pjaxScript = fs.readFileSync(path.join(__dirname, "static-injections", "pjax.js")).toString();
-                    data = data.replace('"REPLACE_WITH_PJAX_INJECTION";', pjaxScript);
-                } else {
-                    data = data.replace('"REPLACE_WITH_PJAX_INJECTION";', "");
-                }
-
-                if (this.config.serviceWorker) {
-                    const swScript = fs.readFileSync(path.join(__dirname, "static-injections", "service-worker.js")).toString();
-                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', swScript);
-                } else if (this.config.pjax) {
-                    const noServiceWorkerScript = fs.readFileSync(path.join(__dirname, "static-injections", "no-service-worker.js")).toString();
-                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', noServiceWorkerScript);
-                } else {
-                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', "");
-                }
-
-                fs.writeFile(runtimeFile, data, error => {
-                    if (error) {
-                        reject(error);
-                    }
-                    resolve();
-                });
-            });
-        });
-    }
-
-    injectDjinnjax() {
-        return new Promise((resolve, reject) => {
-            const djinnjax = path.resolve(cwd, "./node_modules/djinnjax/dist/djinnjax.js");
-            if (!fs.existsSync(djinnjax)) {
-                reject("The djinnjax package is not installed. Disable pjax or run 'npm i -S djinnjax'");
-            } else {
-                glob(`${path.resolve(djinnjax, "../")}/*.js`, (error, files) => {
-                    if (error) {
-                        reject(error);
-                    }
-                    const tempDir = path.join(__dirname, "injections");
-                    let count = 0;
-                    for (let i = 0; i < files.length; i++) {
-                        const filename = files[i].replace(/(.*[\/\\])/g, "");
-                        fs.copyFile(files[i], `${tempDir}/${filename}`, error => {
-                            if (error) {
-                                reject(error);
-                            }
-                            count++;
-                            if (count === files.length) {
-                                resolve();
-                            }
-                        });
-                    }
-                });
-            }
-        });
-    }
-
-    async dependencyInjection() {
-        try {
-            if (this.config.pjax) {
-                await this.injectDjinnjax();
-            }
-        } catch (error) {
-            throw error;
-        }
     }
 
     scrubScripts() {
@@ -400,25 +254,10 @@ class DjinnJS {
     cleanup() {
         return new Promise(resolve => {
             const tempPath = path.join(__dirname, "temp");
-            const injectionsPath = path.join(__dirname, "injections");
             if (fs.existsSync(tempPath)) {
                 rimraf.sync(tempPath);
             }
-            if (fs.existsSync(injectionsPath)) {
-                rimraf.sync(injectionsPath);
-            }
             resolve();
-        });
-    }
-
-    createInjectionsDirectory() {
-        return new Promise((resolve, reject) => {
-            fs.mkdir(path.join(__dirname, "injections"), error => {
-                if (error) {
-                    reject(error);
-                }
-                resolve();
-            });
         });
     }