+++ cli/static-injections/no-service-worker.js
@@ -1,2 +1,0 @@
-const event = new CustomEvent("pjax:init");
-document.dispatchEvent(event);