+++ src/env.ts
@@ -1,4 +1,4 @@
-type DOMState = "soft-loading" | "hard-loading" | "idling" | "page-loading" | "page-loading-complete";
+type DOMState = "loading" | "idling";
 type NetworkType = "4g" | "3g" | "2g" | "slow-2g";
 type Browser = "chrome" | "safari" | "edge" | "chromium-edge" | "ie" | "firefox" | "unknown" | "opera";
 
@@ -19,7 +19,7 @@ class Env {
         // Automatically removing 2 since DjinnJS has 2 critical web workers
         this.threadPool = this.cpu - 2;
         this.connection = "4g";
-        this.domState = "hard-loading";
+        this.domState = "loading";
         this.dataSaver = false;
         this.browser = "unknown";
         this.setBrowser();
@@ -68,7 +68,7 @@ class Env {
             }
         }
 
-        if (this.tickets.length === 0 && this.domState === "soft-loading") {
+        if (this.tickets.length === 0 && this.domState === "loading") {
             this.domState = "idling";
             document.documentElement.setAttribute("state", this.domState);
         }
@@ -79,32 +79,13 @@ class Env {
      * @returns a ticket `string` that is required to stop the loading state.
      */
     public startLoading(): string {
-        if (this.domState !== "hard-loading") {
-            this.domState = "soft-loading";
+        if (this.domState !== "loading") {
+            this.domState = "loading";
             document.documentElement.setAttribute("state", this.domState);
         }
         const ticket = this.uid();
         this.tickets.push(ticket);
         return ticket;
-    }
-
-    public startPageTransition(): void {
-        this.domState = "page-loading";
-        document.documentElement.setAttribute("state", this.domState);
-    }
-
-    public endPageTransition(): void {
-        this.domState = "page-loading-complete";
-        document.documentElement.setAttribute("state", this.domState);
-        setTimeout(() => {
-            if (this.tickets.length) {
-                this.domState = "soft-loading";
-                document.documentElement.setAttribute("state", this.domState);
-            } else {
-                this.domState = "idling";
-                document.documentElement.setAttribute("state", this.domState);
-            }
-        }, 600);
     }
 
     /**